Improve usage and refactor for new CTD file
Output directories and the locations of the data directories were
hardcoded to locations on the internal ECCC supercomputers. This
has been changed so that that these are now runtime-configurable
via argparse. Additionally, these are now transformed to
pathlib.Path objects directly and some of the latter uses when
constructing paths to subdirectories and files now use Path
semantics.

To help new users of the repository, the README.md has been
updated to describe how to install dependencies (using the new
requirements.txt), download the training  data, and how to run
the script.

An initial refactor of the `load_ctd_data` function has been done
to support the new Line P CTD data file that contains the entirety
of the dataset. Problems linger associated with constructing the
stations.